pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "videoapi"
version = "0.1.0"
description = "A lightweight, configurable video streaming and recording API"
authors = [{name = "Carmelo Gonzales", email = "carmelojsg@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "opencv-python>=4.7.0",
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[project.scripts]
videoapi = "videoapi.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
videoapi = ["configs/*.yaml"]
